requirements.txt
astroid==2.3.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
cryptos==1.36
dj-database-url==0.5.0
Django==2.2.5
django-db==0.0.7
django-filter==2.2.0
django-friendship==1.8.1
django-heroku==0.3.1
django-oauth-toolkit==1.2.0
django-qr-code==1.1.0
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
Markdown==3.1.1
mccabe==0.6.1
mysqlclient==1.4.4
oauthlib==3.1.0
Pillow==6.1.0
psycopg2==2.8.4
psycopg2-binary==2.8.3
pylint==2.4.1
PyMySQL==0.9.3
pytz==2019.2
qrcode==6.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.2.0
urllib3==1.25.6
whitenoise==4.1.4
wrapt==1.11.2